change field type

+++ goods/models.py
@@ -6,6 +6,7 @@ from django.core.validators import MinValueValidator
 from django.db import models
 from django.db.models import CharField
 from django.db.models import FloatField
+from django.db.models import IntegerField
 from django.db.models import UUIDField
 from django.db.models import ForeignKey
 from django.db.models import DateTimeField
@@ -38,7 +39,7 @@ class ShopUnitOffer(models.Model):
     id = UUIDField(primary_key=True, editable=False)
     name = CharField(max_length=255, null=False)
     parent_id = ForeignKey('ShopUnitCategory', on_delete=models.CASCADE, null=True, blank=True)
-    price = FloatField(null=False, validators=[MinValueValidator(0)])
+    price = IntegerField(null=False, validators=[MinValueValidator(0)])
     type = CharField(max_length=20, default=ShopUnitType.OFFER, null=False, editable=False)
     date = DateTimeField()
 
@@ -47,7 +48,7 @@ class ShopUnitCategory(models.Model):
     id = UUIDField(primary_key=True, editable=False)
     name = CharField(max_length=255, null=False)
     parent_id = ForeignKey('ShopUnitCategory', on_delete=models.CASCADE, null=True, blank=True)
-    price = FloatField(null=True, blank=True, default=None, editable=False, validators=[equal_null])
+    price = IntegerField(null=True, blank=True, default=None, editable=False, validators=[equal_null])
     type = CharField(max_length=20, default=ShopUnitType.CATEGORY, null=False, editable=False)
     date = DateTimeField()
